requirements.txt
# This file tells the deployment toolkit what libraries your agent needs

bedrock-agentcore
strands-agents